Add parameters to generate_data and default envs
