+++ experiment_params/default_environments/spring.yaml
@@ -1,0 +1,4 @@
+environment:
+  name: "Spring"
+  mass: 0.5
+  elastic_cst: 2.0
