+++ experiment_params/default_environments/two_bodies.yaml
@@ -1,0 +1,5 @@
+environment:
+  name: "NObjectGravity"
+  mass: [0.5, 0.5]
+  g: 1.0
+  orbit_noise: 0.1
